+++ CHANGELOG.md
@@ -28,6 +28,8 @@
   - com.dwarfeng.familyhelper.clannad.stack.bean.entity.Certificate。
   - com.dwarfeng.familyhelper.clannad.stack.bean.entity.CertificateFileInfo。
   - com.dwarfeng.familyhelper.clannad.stack.bean.entity.Poce。
+  - com.dwarfeng.familyhelper.clannad.stack.bean.entity.NotifySetting。
+  - com.dwarfeng.familyhelper.clannad.stack.bean.entity.NotifyTopic。
 
 - 优化操作服务验证环节的代码结构。